Improved tmux 1.5 integration; no more window renaming
Currently, screen.vim uses a magic window title to determine which
window contains the shell pane. This is fragile, mildly annoying, and
also not necessary in Tmux 1.5 (released July 2011).

`tmux list-panes` gained an `-a` flag in the latest version, which lists
all windows along with their unique pane ids. This list can be grepped
to determine the window index of a pane if the unique pane id is known.

First we remove or noop all instances in which we would have set a tmux
window title. Then, we extract and consolidate all calls to select-pane
or select-window into a new method s:screenTmux:focusPane() [1], which
implements the new behavior.

This commit breaks compatibility with tmux 1.4.

[1]: I did not overload s:screenTmux:focus(), because that would have
     violated the arity parity with s:screenGnuScreen:focus()